Initial commit of non-workspace fork
This is mainly to appease Cargo and the trait resolution. Essentially
there are a few issues that occur when specifying a Git dependency of
a fork of a workspace parent package.

1. The git dependency ends up overriding all the children workspace
   members, not just the one that was specified.
2. Due to the usage of `path` members, this means that the dependencies
   of the workspace member get resolved (and duplicated) to the git
   dependency
3. Due to the implementation of specific traits, you end up
   using a trait from crate A but the trait is implemented for
   duplicate crate A

There's a lot of issues that occur cause of this, including a
compilation error and bloated up binaries due to double the dependency
tree.

It'd be nice if the patch was accepted upstream, but for now this is
the best I can do.